make usage report form can use again. move usage report from common/compound into report folder.
