+++ client/src/app/views/common/compounds/index.jsx
@@ -1,3 +1,4 @@
 export { default as MenuButton } from './menu-button'
 export { default as Project } from './project'
-export { default as UsageReportChart } from './usage-report-chart'
+export { default as UsageReportChart } from './report/usage-report-chart'
+export { default as UsageForm } from './report/usage-form'
