+++ client/src/app/views/common/compounds/report/usage-report-chart.jsx
